+++ casia/src/actions/index.js
@@ -89,3 +89,13 @@ export const unarchiveNote = ID => {
     }
   };
 };
+
+export const AddTask = task => {
+  return {
+    type: 'ADD_TASK',
+    payload: {
+      listValue: task.value,
+      listIcon: task.image
+    }
+  };
+};